Added Loading animation to productScreen Added snackBar to the productScreen

+++ app/src/main/java/com/example/shopapp/feature_store/presentation/cart/CartScreen.kt
@@ -6,13 +6,21 @@ import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.example.shopapp.feature_store.data.entity.Cart
 import com.example.shopapp.feature_store.presentation.components.CartItem
+import com.example.shopapp.feature_store.presentation.components.SearchBar
 
 
 @Composable
@@ -20,6 +28,10 @@ fun CartScreen(
     navController: NavController,
     viewModel: CartViewModel = hiltViewModel()
 ) {
+
+    val searchState = remember {
+        mutableStateOf(TextFieldValue(""))
+    }
     Scaffold(
         topBar = {
             TopAppBar(
@@ -37,27 +49,40 @@ fun CartScreen(
         }
     ) { padding->
         val cartState = viewModel.cartState.value
-        Box(modifier = Modifier
+        Column(modifier = Modifier
             .fillMaxSize()
             .padding(padding))
         {
-            LazyColumn()
+            Spacer(modifier = Modifier.height(8.dp))
+
+            if(cartState.productWithCart.isEmpty())
             {
-                items(cartState.productWithCart.size) { idx ->
-                    CartItem(
-                        productWithCart = cartState.productWithCart[idx],
+                Column(modifier = Modifier.fillMaxSize(), verticalArrangement = Arrangement.Center, horizontalAlignment = Alignment.CenterHorizontally) {
+                    Text("Empty Cart!!", color = Color(0xFFDA2079), fontWeight = FontWeight.Bold, fontSize = 40.sp)
+                }
+            }else{
+                SearchBar(searchState,onSearchAction = {
+                    viewModel.onEvent(CartEvent.Search(it))
+                })
+                Spacer(modifier = Modifier.height(8.dp))
+                LazyColumn()
+                {
+                    items(cartState.productWithCart.size) { idx ->
+                        CartItem(
+                            productWithCart = cartState.productWithCart[idx],
                             modifier= Modifier
                                 .fillMaxWidth()
                                 .height(150.dp)
                                 .padding(4.dp),
-                        onDeleteClicked = {
-                                          viewModel.onEvent(CartEvent.DeleteCart(it))
-                        },
-                        onButtonClicked = {
-                            val cart=cartState.productWithCart[idx].cart
-                            viewModel.onEvent(CartEvent.UpdateCart(Cart(cart.id,cart.productId,it)))
-                        }
-                    )
+                            onDeleteClicked = {
+                                viewModel.onEvent(CartEvent.DeleteCart(it))
+                            },
+                            onButtonClicked = {
+                                val cart=cartState.productWithCart[idx].cart
+                                viewModel.onEvent(CartEvent.UpdateCart(Cart(cart.id,cart.productId,it)))
+                            }
+                        )
+                    }
                 }
             }
         }
